backend/requirements.txt
supabase==2.3.4
gotrue==1.3.1
python-dotenv==1.0.0
openai==1.12.0
crewai==0.5.0
pydantic==2.6.1
pydantic-settings==2.2.1
asyncpg==0.29.0
pgvector==0.2.4
httpx==0.25.2
typer==0.9.0
rich==13.7.0
pytest==8.0.0
pytest-asyncio==0.23.5
